+++ lib/IamGroup.ts
@@ -1,24 +1,13 @@
 import { Construct } from 'constructs';
 import { IGroup, Group, IUser } from 'aws-cdk-lib/aws-iam';
-import { ICdkGroup, CdkUsers } from '../config/parameters';
+import { ICdkGroup, CdkUsers, CdkGroups } from '../config/parameters';
+import { IamRole } from './IamRole';
 
 export class IamGroup {
   private readonly scope: Construct;
 
   constructor(scope: Construct) {
     this.scope = scope; 
-  }
-  
-  public Get(params: ICdkGroup): IGroup {
-    const { groupName } = params;
-    try {
-      return Group.fromGroupName(this.scope, groupName, groupName) as Group;
-    } catch (error) {
-      if (error instanceof Error && error.message.includes('not found')) {
-        return this.Set(params) as Group;
-      }
-      throw error;
-    }
   }
 
   public Set(params: ICdkGroup): Group {
@@ -48,7 +37,15 @@ export class IamGroup {
     });
   }
 
-  public GetGroup(groupName: string): IGroup {
+  public AddGroupToAssignedRoles(Group: Group): void {
+    const cdkGroup = CdkGroups.find((group) => group.groupName === Group.node.id);
+    return cdkGroup?.roles?.forEach((role) => {
+      const awsRole = new IamRole(this.scope).GetIamRole(role);
+      awsRole.grantAssumeRole(Group);
+    });
+  }
+
+  public GetIamGroup(groupName: string): IGroup {
     return this.scope.node.tryFindChild(groupName) as Group ;
   }
 }