+++ lib/IamRole.ts
@@ -1,8 +1,9 @@
 import { Construct } from 'constructs';
 import { Stack } from 'aws-cdk-lib';
-import { IPrincipal, Role, AccountPrincipal } from 'aws-cdk-lib/aws-iam';
-import { CdkRoles, ICdkRole } from '../config/parameters';
-import { IamGroup } from './IamGroup';
+import { IPrincipal, IRole, AccountPrincipal, ManagedPolicy, Role } from 'aws-cdk-lib/aws-iam';
+import { CdkRoles, CdkUsers, ICdkRole, ICdkUser } from '../config/parameters';
+import { IamPolicy } from './IamPolicy';
+import { IamUser } from './IamUser';
 
 export class IamRole {
   private readonly scope: Construct;
@@ -11,30 +12,39 @@ export class IamRole {
     this.scope = scope;
   }
 
-  public Set(params: ICdkRole): Role {
-    const { path, roleName } = params;
-    const assumedBy: IPrincipal = new AccountPrincipal(Stack.of(this.scope).account);
-    return this.GetRole(roleName) ??
-      new Role(this.scope, roleName, {
-        roleName,
-        path,
-        assumedBy,
-      });
-  }
-
-  public AddRoleToAssignedGroups(Role: Role): void {
-    const cdkRole = CdkRoles.find((role) => role.roleName === Role.node.id);
-    return cdkRole?.allowedGroups.forEach((group) => {
-      const awsGroup = new IamGroup(this.scope).GetGroup(group);
-      Role.grantAssumeRole(awsGroup);
+  public AddRoleToAssignedPolicies(Role: IRole): void {
+    const cdkRole = this.GetCdkRole(Role);
+    return cdkRole?.policies?.forEach((policy) => {
+      const awsPolicy = new IamPolicy(this.scope).GetIamPolicy(policy) as ManagedPolicy;
+      awsPolicy.attachToRole(Role);
     });
   }
 
-  public GetRole(roleName: string): Role {
+  public AddRoleToAssignedUsers(Role: IRole): void {
+    const cdkUsers: ICdkUser[] = CdkUsers.filter(user => user.roles?.includes(Role.node.id as string));
+    return cdkUsers.forEach(user => Role.grantAssumeRole(new IamUser(this.scope).GetIamUser(user.userName)));
+  }
+
+  private GetCdkRole(Role: IRole): ICdkRole {
+    return CdkRoles.find((role) => role.roleName === Role.node.id) as ICdkRole;
+  }
+
+  public GetIamRole(roleName: string): Role {
     try {
       return this.scope.node.tryFindChild(roleName) as Role;
     } catch {
       return Role.fromRoleName(this.scope, roleName, roleName) as Role;
     }
+  }
+  
+  public Set(params: ICdkRole): Role {
+    const { path, roleName } = params;
+    const assumedBy: IPrincipal = new AccountPrincipal(Stack.of(this.scope).account);
+    return this.GetIamRole(roleName) ??
+      new Role(this.scope, roleName, {
+        roleName,
+        path,
+        assumedBy,
+      });
   }
 }